group work

+++ pré-traitement/Rotations/rotationMAN.h
@@ -10,6 +10,6 @@
 
 struct Tuple{int x; int y;};
 struct Tuple shear3(double angle,int x, int y);
-SDL_Surface* ManualRotation(SDL_Surface* image_surface , int degrees);
+SDL_Surface* ManualRotation(SDL_Surface* image_surface , double degrees);
 
 #endif
